+++ butler/src/main/java/anicare/butler/ButlerApplication.java
@@ -2,8 +2,12 @@ package anicare.butler;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EntityScan("anicare.logic")
+@EnableJpaRepositories("anicare.logic")
 public class ButlerApplication {
 
 	public static void main(String[] args) {
