--- butler/src/main/java/anicare/controller/UserController.java
@@ -1,4 +1,0 @@
-package anicare.controller;
-
-public class UserController {
-}
