--- butler/src/main/java/anicare/dao/user/UserDao.java
@@ -1,28 +1,0 @@
-package anicare.dao.user;
-
-import lombok.*;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-@Builder
-public class UserDao {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String name; // 본인 이름
-
-    private  String email; // 아이디
-
-    private  String password; // 비밀번호
-
-    private String animal;
-}
