+++ butler/src/main/java/anicare/logic/service/UserService.java
@@ -1,0 +1,26 @@
+package anicare.logic.service;
+
+import anicare.logic.dto.user.UserSignupDto;
+import anicare.logic.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class UserService {
+
+    @Autowired
+    public UserRepository userRepository;
+
+    /*
+    중복이메일 확인
+     */
+
+    @Transactional
+    public Boolean CheckEmail(String email){
+       if(userRepository.findByEmail(email)){
+           return false;
+       }
+       return true;
+    }
+}
