--- butler/src/main/java/anicare/service/UserService.java
@@ -1,7 +1,0 @@
-package anicare.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-}
